requirements.txt
flet
mipa
win11toast
python-dotenv